requirements.txt
crowdin-api-client